Add bootstrap query script
During the migration from an APIC M1/M2/M3/L1/L2/L3 to M4/L4, you may experience long wait times during the initial commission/bootstrap of the new APIC(s). During this time, you are not presented with any progress status in the GUI. This can leave one wondering whether any progress is being made at all. The purpose of this program is to provide an initial bootstrap progress status using an API call.

The program performs a GET request on an APIC using the initial cluster bootstrap API (/api/workflows/v1/cluster/bootstrap), and provides a JSON response with detailed information on the progress of each of the below bootstrap elements:

1. Overall progress status
2. Cluster config validation
3. Config objects push
4. Mgmt network bringup
5. Bootstrap progress status
6. Operation performed